--- src/components/Home/Header/index.stories.tsx
@@ -1,11 +1,0 @@
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { Header } from '.';
-
-export default {
-  title: 'Component/Header',
-  component: Header,
-} as ComponentMeta<typeof Header>;
-
-const Template: ComponentStory<typeof Header> = () => <Header />;
-
-export const LoggedOut = Template.bind({});
