+++ src/components/Home/Paragraph/styles.ts
@@ -3,6 +3,7 @@ import styled from '@emotion/styled';
 export const Paragraph = styled.section`
   max-width: 1200px;
   width: 100%;
+  height: 40px;
   display: flex;
   justify-content: space-between;
   margin: auto;
